handle turned off bluetooth during onboarding
