+++ src/xcode/ENA/ENA/Source/Services/ExposureSubmissionService.swift
@@ -223,7 +223,7 @@ class ENAExposureSubmissionService: ExposureSubmissionService {
 
 		if let exposureNotificationError = error as? ExposureNotificationError {
 			switch exposureNotificationError {
-			case .exposureNotificationRequired, .exposureNotificationAuthorization, .exposureNotificationUnavailable, .bluetoothOff:
+			case .exposureNotificationRequired, .exposureNotificationAuthorization, .exposureNotificationUnavailable:
 				return .enNotEnabled
 			}
 		}
